pyproject.toml
[project]
name = "hippobox"
version = "0.1.0"
description = "HippoBox MCP + Knowledge Store Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "mcp[cli]>=1.17.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.6.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.1",
    "black>=25.11.0",
    "isort>=7.0.0",
    "flake8>=7.3.0",
    "fastapi-mcp>=0.4.0",
    "openai>=2.8.1",
    "langchain-openai>=1.0.3",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "redis[hiredis]>=5.0",
    "bcrypt>=5.0.0",
    "argon2-cffi>=25.1.0",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
